Improved Layout
While the controls are a bit more nicely laid out, the <pre>
tag containing the actual cowsay output is oversizing the
div that I tried to wrap around it. Also, the bottom of the
saved output is covered by the footer, and it's not possible
to scroll the page up anymore.